requirements.txt
cryptography~=42.0.5
pillow~=10.2.0
colorama~=0.4.6